Added Dashboard Consulting page with images.
Note: I included images probably in a non-ideal way. I think in Gatsby it could be much easier to include images, for example with GraphQL queries and use responsive sizes and all that jazz. Howver I currently don't know how to use that and don't have the time yet to learn. It's more important to just get the site out there so I can stop using the ThemeCo hosting.