Fix that removes two React warnings from the console
When developing locally there were two React errors. One about a PropType and another about list items with duplicate keys.

Aparently the GatsbyJS starter blog that I based mine off of needs the "path" property in markdown blog posts. So in the current setup path is determined by the folder name. Then when the list of blog links is created the "path" property is used. I think this could be refined so that only one source of truth is used.