Removed a deprecated React createClass(), used ES2015 class instead
I was trying to get rid of warnings in the console when running the site locally. createClass() is deprecated in the next v16 of React. However this unfortunetly did not get rid of the warnings. I suspect that GatsbyJS itself (or some dependencies) are still using createClass() so it is mostly out of my reach.