Attempting to get rid of all React PropType deprecation warning
React.PropType is getting deprecated in React v16. There is a warning message in the Chrome console I was trying to get rid of. However as with the createClass() warnings it is not my own code at fault. GatsbyJS internals (or its dependencies) use React.PropTypes which propagates as a warning. I will just have to wait until it is fixed in GatsbyJS.