Fixes issue #18: robots.txt breaks the master deployment branch
Instead of a robots.txt file that I delete, I create the file on the fly if the deployment is on the dev branch.

Sidenote: the robots.txt file that I had was missing the / character in Dissalow. That means the file was useless as it allowed all content.

See: https://github.com/jonrh/jonrh.is/issues/18