Updated to latest dependencies
The dependency gatsby-remark-prismjs is not latest. At the time of writing it was 2.0.3 or 2.0.3-3, I'm picking 1.2.24. I tried the newest but it wouldn't compile. Didn't dig too much into it but I suspect it's intended for the GatsbyJS v2 alpha release. I couldn't find any changelog or anything clarifying it. Another GatsbyJS fault.