Fixed an error in Rollbar config
The variable `string` was not defined. This somehow resulted in an error that propagated to GatsbyJS not being able to build the files `index.html`.

Really odd behaviour. I would have expected GatsbyJS to print out some sort of error to the console in the event of a build error or a warning.