Added underline to header navigation links
Before this change the About/Blog/Dashboard links got an underline once hovered over. Changed it to be consistent with other links of the site. That is, an underline is displayed and when hovered the underline disappears reaffirming it's a link. I think this makes the site slightly more accessible visually.