Added pictures of Tímavera dashboard and app
The screenshots are from the frontpage of timavera.is. For convenience the screenshot of the app is 3 screenshots appended together.

The screenshots are in Icelandic. Ideally I'd want them in English but I haven't created English screenshots yet for Tímavera. Once I will do that I aim to update these.