Added /portfolio page and added a link to it in the main navigation
This new /portfolio page is a copy of dashboard-consulting.js that I then modified. I have also replaced the link to "Dashboard Consulting" in the main header navigation. Instead I now link to Portfolio.

The text copy could use some work, I would consider it to be in a preliminary state, especially the sections on Tímavera and Activity Stream. The text is basically a copy paste from my resume for now.