CircleCI deployments: move robots.txt generation to a separate step
Moved the creation of the robots.txt file to a separate step in the CircleCI job. Additionally now we create a robots.txt file also on production (jonrh.is). Previously we just left it empty. From the top of my head I think that the default behaviour is that if a robots.txt file is missing it's assumed that the entire site can be indexed so it shouldn't have mattered much.