Upgraded Gatsby from v1.9.270 to v2.24.14
I could no longer develop locally on my machine. Something to do with the dependency of sharp and node gyp. I've had this issue multiple times before so I took the decision to take a different approach to using GatsbyJS for my site.

Currently I use markdown for the blog posts but after about 2.5 years I have found it to be cumbersome. It complicates the setup by requiring a great deal of plugins that I have found to be unreliable. Recently I converted the website jonogmarteinn.is to a Gatsby site but didn't use any Markdown. Just React components and JavaScript. It was refreshingly simple and fun and the plan is to do the same for this site.

So as of this commit all the blog posts are inaccessible because they have yet to be converted from Markdown into React components in JavaScript. That will come later.

Use the default `html.js` provided by GatsbyJS by deleting `src/html.js`. Since `src/html.js` (my overwrite) was deleted Rollbar error reporting is no longer being included. So for now the Rollbar component code was deleted. I will probably add that back later.