Renamed all blog post folders, convert all markdown files to JavaScript
Renamed all blog post folders from the following pattern:

YYYY-MM-DD---blog-post-url/

to:

blog-post-url/

Additionally, all index.md files were renamed index.js. The content is still all markdown so as of this commit none of the files work. A conversion to JavaScript will come in a later commit.

This is deliberately done in two separate commits so the full history of the documents will be retained.