Relocated files to match the expected Next.js file structure
The renaming and relocation of these files and folders is to facilitate a migration from GatsbyJS to Next.js v10.

Renamed all blog post folders. Removed the "YYYY-MM-DD---" prefix. This prefix was a legacy naming scheme from Gatsby v0 beta from 3 years ago. As I recall it was required for Gatsby to know what files should be considered Markdown blog posts. There is now no longer any need for this and in order to retain the same URL structure the folders will need to be renamed as they have.