Deleted CircleCI config (AWS + S3 + CF), Vercel will be used instead
This CircleCI config file was used to deploy the website to AWS. Static files were stored in S3 and then globally distributed with CloudFront. I will be using Vercel hosting now that I have Next.js. I have used it in a few projects before and it's a terrific combination.

As far as I know, Vercel actually used S3 and CloudFront under the hood but now I do not have to worry about it.