Converted all markdown blog posts to MDX, relocated imgs to public/
Post.js is a template file/component for each and every blogpost. Originally designed to be used by MDX files.

To show images I use the newly introduced next/image component. I think it works similarly to the resharp image component of Gatsby. It creates an <img> tag with a sourceset of the images in various sizes. What is new with this component is that width and height has to be defined. I just use the original dimensions of the image so the aspect ratio stays the same. The style layout already makes sure that images don't go out of bounds for large source images.

The next/image component expects images to be located in the public/ folder. I couldn't get it to work otherwise. Therefore all blog post images have been relocated to public/images/name-of-blogpost-folder/.

Deleted the markdown front matter of ever post as it is no longer relevant. Converted it to <Post title="Blog post title" date="October 28, 2020" /> instead.