Added lang="en" attribute to <html> tag
Lighthouse pointed out that specifying a language on the HTML element is better for accessibility. Adding this config will have the following effect on all pages.

Before:
<html>

After:
<html lang="en">

This is a bit of a roundabout way of implementing this in Next.js but the best option I have found so far.

See:
https://github.com/vercel/next.js/issues/9160#issuecomment-752713231