Removed moment as a dependency greatly reducing payload size
Moment was only being used for automatically calculating how many years of experience I have in frontend engineering. In retrospect it was a huge waste to pull in the entire Moment library for such simple calculations. The Moment calculations have now been replaced with this:

new Date().getUTCFullYear() - 2014;

It's a bit more crude but does the job well enough. I think this change shaved off 60-167KB. It's a bit of a fuzzy calculation. I thought Moment was only supposed to be something like 18KB minified. However when comparing before the change the Chrome Network tab is saying that 461KB were transferred before and 295KB after the change. In any case I think this should speed the site up nicely.

Another benefit of getting rid of Moment is that the library has been a steady source of security notifications on the GitHub repository. Hopefully there will be fewer of those now.