Replaced profile picture with a dithered version, 113KB -> 5KB
Replaced to an alternative profile picture. A picture taken in front of Skógarfoss. Removed the background and saved as a 2 color dithered image. The image in use before was 113KB. The new dithered image is just under 5KB.

Removed the profile picture circular form, border and shadow. Looks better without it with a dithered image.