Add word wrapping in code blocks to fix broken layout
When writing code in markdown files such as

```
code here
```

it would get converted into

<pre>
  <code>
    code here
  </code
</pre>

after a transform. The <pre> tag is for preformatted text. That is it leases it untouched. That is not exactly desired when I display code on the site. To make it fit within the content width this change now wraps on whitespace when needed. Does not look great but an improvement on how it was before. Especially on really long lines.

See: https://stackoverflow.com/a/6098078