Switch to a lighter orange colour, darker background and text color
Now that the site is going to be going all in on dark mode I can go for the original orange colour I had in mind: #ff7b00. It is the colour I used in the Bespoke Dashboards logo. I recall having given some thought to that colour so I am going to use it for now. The previous orange colour was more like a dark red/orange.

Made new versions of the dithered images. I could have used a CSS `filter: hue-rotate(25deg)` or something like that to shift the colour of the existing images. But since I have not converted that many images yet I think a better way forward is to just use #ff7b00 as the colour and redo the images.

Made the background slightly darker and made the white text slightly more grey. Think it looks a bit better.