Added a Dither component, refactored to use it, renamed & deleted files
Added a recipe for how to generate the dithered images. Using Photoshop for now. Would be an interesting project to use a command line tool in the future.

Deleted some of the original dithering images. Renamed few dithered images to be consistent with the "abc-dithered.png" file name convention.

Added a new component Dither. It displays a dithered image with a link to the original full coloured image. It relies on a file name convention. So if there is an image named a.jpg then there should be an image a-dither.png.

Added dithered images for the React Native 2017 conference blog post.