Renamed /public/images/consulting/ to /public/images/portfolio/, etc
Moved the screenshots of jonogmarteinn.is and malco.is to the newly renamed portfolio/ image folder. They belong there the best I think.

Note that this refactor breaks a whole bunch of images on the portfolio page. In the next commit it will be fixed and refactored into dithered images.