Improve dithered images, added an option for .webp option, etc
Added the `unoptimized={true}` flag to dithered images. By default the Next.js Image component uses quality={75}. The effect of that was that if I created for example an optimised dithered PNG version of an image that was say 25Kb Next.js Image would re-save that PNG as a WebP and the resulting file would look worse (lossy compression) and be about twice the size (50Kb).

This prompted me to dig a little deeper. Turns out Safari has added support for WebP as of iOS 14. WebP achieves better compression than PNGs.

In this commit I added a dark dithered version of the 1276x990 pixel screenshot in one blog post:

Original: 251.037 bytes
PNG: 22.885 bytes
WebP: 14.840 bytes

A monochrome dithered PNG is about 9% of the size of the original screenshot. The WebP is the exact same content (lossless) as the PNG but is 65% of the size of the PNG or 6% of the original. Total savings of about 8Kb for just one image. I would imagine the difference is even bigger for larger screenshots.

So now the dithered images have a new option `webp`. If true it signals that I have made an optimized .webp version of the image that should be used instead. With time I aim to create a .webp version of all images and then I can stop using PNG files for dithering.