Fixed profile picture for browsers that do not support .webp
My Mac mini is running macOS Catalina. WebP support was added in the next version Big Sur. This means the profile image was not showing in Safari.

Fixed it by using the <picture> element. If the browser supports .webp it uses it otherwise it will fall back on a .png version.