Added a enforced 50-160 character meta description to all posts for SEO
All posts must now have have a <meta> description that summarises in 50-160 characters the content of the post.

If the meta description is missing, smaller than 50 characters or larger than 160 characters an error is thrown which halts the deployment of new versions.

From what I could tell by a quick glance the range of 50-160 characters was recommended. It can technically of course be smaller or larger but from what I could tell you would increase the risk of search engines not using the meta description at all or that it would be truncated.

Added a meta description to all existing posts. It was a fair bit of work so the descriptions themselves could be improved.