Fixed a reduced SEO Lighthouse score, tap on <a> inside <li> too small
This fixes the "Tap targets are not sized appropriately" SEO penalty score in a Lighthouse tests. In a list of links such as on the portfolio page it was too hard to press links because their tap areas overlapped. It was actually okay but a bit error prone.

Adding a top and bottom margin to all <li> elements fixes that. I find the style to be good in general. It looks better for multiline <li> elements. Creates a more distinct separation between items.

The `margin: 0;` on the `nav li` CSS selector for the navigation is to counteract the change. It is a special case with the list items laid horizontally so not subject to tap area overlap. The extra top and bottom margin threw off the look of the header navigation. This fixes that in a bit of a hacky way.

See:
https://web.dev/tap-targets/