Fixed a Google Search Console mobile usability warnings
On the custom Totalfinder post there was a path that did not wrap. This resulted in content wider than screen warning which I suspect results in the other two because on Chrome mobile the whole content gets smaller. The errors:

+ Text too small to read
+ Clickable elements too close together
+ Content wider than screen

https://support.google.com/webmasters/answer/9063469#touch_elements_too_close

I essentially added a hacky <div> with a style around the list. The magic trick was the CSS attribute: `overflow-wrap: break-word;`. That wordwraps long words or lines like these paths.

See: https://stackoverflow.com/a/3059128