Fixed skewed images
After updating to Next.js v13 the next/image component switched to the new image component that had breaking changes. The warning message in the browser console:

[Warning] Image with src "/img.png" has either width or height modified, but not the other. If you use CSS to change the size of your image, also include the styles 'width: "auto"' or 'height: "auto"' to maintain the aspect ratio.

Because the existing CSS style on img had `max-width: 100%` (to keep in line with content witdh) most images were displayed in the wrong aspect ratio.