Fixed missing space between # anchor link and heading text
Headings on the portfolio page have a clickable "#" anchor to make it easier to click to and copy a direct link to a specific heading. Before this change there was no spacing between the # and the heading. Example:

#Tímavera

After the change:

# Tímavera

Also removed the underline on the link element, from context it should be clear enough what the purpose is. It is a common design.

Fixed an incorrect href link for Tímavera. Added re-usable components for doing those sort of headings. For now just located on the portfolio page. Do not think they are used anywhere else on the site.