Moved CSS reset code into its own file
When setting up automated Prettier code formatting the CSS reset line was getting auto-formatted. That is minimized CSS that is probably best left untouched although it probably does not hurt to have it be auto-formatted. For now opting for the maximally safe option.

Tried using a `/* prettier-ignore */` but it did not work. Moved the CSS reset line therefore to its own file which can be included in a .prettierignore file.