Added Prettier for automated source code formatting, etc
Added Prettier and Husky lint-staged as dependencies to automatically format source code.

Not all source files are formatted. Prettier seemed to struggle with .mdx files. Formatting some correctly, some inconsistently, and breaking other pages. Disabled for now but might look into enabling later.

Also deliberately opted out of formatting the /tools page. It is mostly HTML content which is easier to work with and parse by allowing some lines to exceed the regular 80ch line length. With automated formatting it was harder to parse when some content was put on new lines but other in the same block not.